requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
filecache==0.81
idna==2.10
python-dotenv==0.15.0
requests==2.25.1
soupsieve==2.2
tqdm==4.59.0
urllib3==1.26.3

pandas~=1.2.3